refactor: move all pages to routes folder
