+++ src/main.tsx
@@ -1,12 +1,15 @@
-import React from 'react';
+import { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
 
 import App from './App.tsx';
 
 import './index.css';
 
 ReactDOM.createRoot(document.querySelector('#root')!).render(
-    <React.StrictMode>
-        <App />
-    </React.StrictMode>,
+    <StrictMode>
+        <BrowserRouter basename="/fe-react-2024">
+            <App />
+        </BrowserRouter>
+    </StrictMode>,
 );
