+++ src/routes/Layout/Layout.module.css
@@ -4,5 +4,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  justify-content: center;
   padding: 0 30px;
 }
