+++ src/routes/ProductsPage/ProductsPage.module.css
@@ -1,0 +1,6 @@
+.productsNotFoundInfo {
+  flex-grow: 1;
+  color: var(--color-text);
+  font-size: 16px;
+  line-height: 20px;
+}
